examples/rabbitmq/requirements.txt
xata
python-dotenv